function that prints all the elements of a listint_t list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,7 +1,6 @@
 #include "lists.h"
 #include <stdio.h>
-#include <string.h>
-#include <stdlib.h>
+
 /**
  * print_listint - prints all the elements of a listint_t list
  * @h: a pointer to the start of the list
@@ -11,7 +10,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-size_t i = 0;
+unsigned int i = 0;
 while (h != NULL)
 {
 printf("%d\n", h->n);
